feat: Added new strings.xml file for Spanish localization
A new file, `strings.xml`, has been added to the `values-es` directory in the `res` folder. This file contains XML code for defining strings in Spanish for localization purposes. The file is empty and lacks a newline at the end.

refactor: Updated app_name string in strings.xml

The `app_name` string in the `strings.xml` file located in the `values` directory has been modified. The `translatable` attribute has been set to `false`, indicating that the string should not be translated. The value of the `app_name` string remains the same, which is "Medical Assistant".

feat: Added a new module named `core/data-test`

A new module named `core/data-test` has been added to the project. This module contains a `README.md` file and an image file representing the dependency graph of the `core/data-test` module.

feat: Added `DoctorId` field to `OfflineFirstUserDataRepositoryTest`